+++ scripts/gc.sh
@@ -35,17 +35,21 @@ cat > $DIRNAME/src/index.vue <<EOF
     <slot></slot>
   </div>
 </template>
+
 <script lang='ts'>
 import { defineComponent } from 'vue'
 
 export default defineComponent({
   name: 'I${NAME}',
-  props: { },
+  props: {
+
+  },
   setup(props) {
     // init here
   },
 })
 </script>
+
 <style>
 </style>
 EOF