+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+#### --- Version 0.4.1
+This version update:
+* Added new element `alert`
+* Fixed form inject/provide error
+
 #### --- Version 0.3.3
 This version update:
 * Changed label style on `form-item` element